fix: only delete residents with delivered parcels

+++ src/main/java/com/smartnotify/resident/usecases/delete/DeleteHorizontalCondoResident.java
@@ -2,6 +2,7 @@ package com.smartnotify.resident.usecases.delete;
 
 import com.smartnotify.condominium.model.CondominiumType;
 import com.smartnotify.config.exception.ResidentNotFoundException;
+import com.smartnotify.parcel.model.DeliveryStatus;
 import com.smartnotify.resident.repository.HorizontalCondoResidentRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -18,6 +19,13 @@ public class DeleteHorizontalCondoResident implements DeleteResident {
         final var resident = repository.findByEmail(email)
                 .orElseThrow(() -> new ResidentNotFoundException("Resident not found"));
 
+        final var hasNotDeliveredParcels = resident.getParcels().stream()
+                .anyMatch(parcel -> parcel.getStatus() == DeliveryStatus.NOT_DELIVERED);
+
+        if (hasNotDeliveredParcels) {
+            throw new IllegalStateException("Error deleting resident " + resident.getName() +
+                    " because of pending parcels. Make sure to deliver all the parcels before deleting the resident.");
+        }
         repository.deleteByEmailAndCondominiumId(resident.getEmail(), condominiumId);
     }
